Refactoring + Bug Fixes
preparation for implementing different types of "best" recipe methods, which take into consideration different criteria for being efficient. potentially in the future we can find a way to weigh the "bests" to have a more generalized "best" that might consider one item to go with a smaller footprint due to how huge others cane be, while on another one goes for reduced power consumption due to the difference in buildings.