Better buildstep + tests
Fleshed out the resulting output of requests made for the entire recipe. Now it shows decimal form of quantities, conditional decimals, and added the item quantites as well. Added one test that tests both the nromalizeSinkPoints and findBestRecipe recursive methods by asking for best screw recipe and getting the default one. (In the future, if there is a tie for best, then we should use another condition to dtermine best between these two. perhaps the best power usage between the two of them)